+++ presto-iceberg/src/main/java/io/prestosql/plugin/iceberg/IcebergMetadata.java
@@ -79,6 +79,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -90,15 +91,12 @@ import static io.prestosql.plugin.hive.HiveColumnHandle.PATH_COLUMN_NAME;
 import static io.prestosql.plugin.hive.HiveColumnHandle.pathColumnHandle;
 import static io.prestosql.plugin.hive.HiveColumnHandle.updateRowIdHandle;
 import static io.prestosql.plugin.hive.HiveErrorCode.HIVE_INVALID_METADATA;
-import static io.prestosql.plugin.hive.HiveMetadata.OWNER;
 import static io.prestosql.plugin.hive.HiveMetadata.PRESTO_QUERY_ID_NAME;
 import static io.prestosql.plugin.hive.HiveMetadata.PRESTO_VERSION_NAME;
 import static io.prestosql.plugin.hive.HiveMetadata.TABLE_COMMENT;
-import static io.prestosql.plugin.hive.HiveMetadata.TABLE_TYPE;
 import static io.prestosql.plugin.hive.HiveSchemaProperties.getLocation;
 import static io.prestosql.plugin.hive.HiveSessionProperties.isCommonViewSupportEnabled;
 import static io.prestosql.plugin.hive.HiveType.HIVE_STRING;
-import static io.prestosql.plugin.hive.HiveUtil.COMMON_VIEW_FLAG;
 import static io.prestosql.plugin.hive.HiveUtil.PRESTO_VIEW_FLAG;
 import static io.prestosql.plugin.hive.HiveUtil.decodeViewData;
 import static io.prestosql.plugin.hive.HiveUtil.encodeViewData;
@@ -602,16 +600,17 @@ public class IcebergMetadata
 
         Map<String, String> properties;
         if (isCommonView) {
-            //String genieJobId = ((FullConnectorSession) session).getSession().getSystemProperty("genie_job_id", String.class);
+            Map buildProps = new HashMap();
+            buildProps.put(PRESTO_VERSION_NAME, prestoVersion);
+            buildProps.put(PRESTO_QUERY_ID_NAME, session.getQueryId());
+            buildProps.put(CommonViewConstants.ENGINE_VERSION, prestoVersion);
+            buildProps.put(CommonViewConstants.GENIE_ID, session.getGenieJobId());
+            if (!replace) {
+                // Add a default table comment when the view is being created
+                buildProps.put(TABLE_COMMENT, "Common View created from Presto");
+            }
             properties = ImmutableMap.<String, String>builder()
-                    .put(TABLE_COMMENT, "Common View created from Presto")
-                    .put(COMMON_VIEW_FLAG, "true")
-                    .put(PRESTO_VERSION_NAME, prestoVersion)
-                    .put(PRESTO_QUERY_ID_NAME, session.getQueryId())
-                    .put(OWNER, session.getUser())
-                    .put(TABLE_TYPE, org.apache.hadoop.hive.metastore.TableType.VIRTUAL_VIEW.name())
-                    .put(CommonViewConstants.ENGINE_VERSION, prestoVersion)
-                    //.put(CommonViewConstants.GENIE_ID, genieJobId)
+                    .putAll(buildProps)
                     .build();
         }
         else {
@@ -620,8 +619,6 @@ public class IcebergMetadata
                     .put(PRESTO_VIEW_FLAG, "true")
                     .put(PRESTO_VERSION_NAME, prestoVersion)
                     .put(PRESTO_QUERY_ID_NAME, session.getQueryId())
-                    .put(OWNER, session.getUser())
-                    .put(TABLE_TYPE, org.apache.hadoop.hive.metastore.TableType.VIRTUAL_VIEW.name())
                     .build();
         }
 
@@ -674,6 +671,19 @@ public class IcebergMetadata
     @Override
     public void dropView(ConnectorSession session, SchemaTableName viewName)
     {
+        Optional<Table> view = metastore.getTable(viewName.getSchemaName(), viewName.getTableName());
+        if (!view.isPresent()) {
+            throw new ViewNotFoundException(viewName);
+        }
+        boolean isCommonView = HiveUtil.isCommonView(view.get());
+        if (!isCommonViewSupportEnabled(session) && isCommonView) {
+            throw new PrestoException(NOT_SUPPORTED, "This connector does not support common views.");
+        }
+
+        if (isCommonView) {
+            Configuration configuration = getConfiguration(session, viewName.getSchemaName());
+            commonViewUtils.dropView(configuration, session.getCatalog(), viewName);
+        }
         try {
             metastore.dropTable(viewName.getSchemaName(), viewName.getTableName(), true);
         }
